[HW3] extend unitest
